LRCI-3783 Move get downstream build protected methods into BaseParentBuild
