+++ modules/test/jenkins-results-parser/src/main/java/com/liferay/jenkins/results/parser/BaseParentBuild.java
@@ -9,9 +9,13 @@ import java.io.UnsupportedEncodingException;
 
 import java.util.ArrayList;
 import java.util.HashMap;
+import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.Objects;
 import java.util.concurrent.Callable;
+
+import org.dom4j.Element;
 
 /**
  * @author Michael Hashimoto
@@ -365,6 +369,67 @@ public abstract class BaseParentBuild extends BaseBuild implements ParentBuild {
 		for (Build downstreamBuild : getDownstreamBuilds(null)) {
 			downstreamBuild.addTimelineData(timelineData);
 		}
+	}
+
+	protected int getDownstreamBuildCountByResult(String result) {
+		List<Build> downstreamBuilds = getDownstreamBuilds(null);
+
+		if (result == null) {
+			return downstreamBuilds.size();
+		}
+
+		int count = 0;
+
+		for (Build downstreamBuild : downstreamBuilds) {
+			String downstreamBuildResult = downstreamBuild.getResult();
+
+			if (Objects.equals(downstreamBuildResult, result)) {
+				count++;
+			}
+		}
+
+		return count;
+	}
+
+	protected Map<Build, Element> getDownstreamBuildMessages(
+		List<Build> downstreamBuilds) {
+
+		List<Callable<Element>> callables = new ArrayList<>();
+
+		for (final Build downstreamBuild : downstreamBuilds) {
+			Callable<Element> callable = new Callable<Element>() {
+
+				public Element call() {
+					return downstreamBuild.getGitHubMessageElement();
+				}
+
+			};
+
+			callables.add(callable);
+		}
+
+		ParallelExecutor<Element> parallelExecutor = new ParallelExecutor<>(
+			callables, getExecutorService());
+
+		List<Element> elements = parallelExecutor.execute();
+
+		Map<Build, Element> elementsMap = new LinkedHashMap<>();
+
+		for (int i = 0; i < elements.size(); i++) {
+			elementsMap.put(downstreamBuilds.get(i), elements.get(i));
+		}
+
+		return elementsMap;
+	}
+
+	protected List<Build> getFailedDownstreamBuilds() {
+		List<Build> failedDownstreamBuilds = new ArrayList<>();
+
+		failedDownstreamBuilds.addAll(getDownstreamBuilds("ABORTED", null));
+		failedDownstreamBuilds.addAll(getDownstreamBuilds("FAILURE", null));
+		failedDownstreamBuilds.addAll(getDownstreamBuilds("UNSTABLE", null));
+
+		return failedDownstreamBuilds;
 	}
 
 }
